@@ -2,13 +2,8 @@
   <Import Project="..\common.props" />
 
   <PropertyGroup>    
-    <TargetFrameworks>netcoreapp2.0;net461;</TargetFrameworks>
+    <TargetFrameworks>netcoreapp2.1;net461;</TargetFrameworks>
   </PropertyGroup>
-
-  <ItemGroup>
-    <None Remove="JsonSerializationSpec.JsonApprovalSpecWithParent.approved.json" />
-    <None Remove="JsonSerializationSpec.JsonApprovalSpecWithSimpleSpan.approved.json" />
-  </ItemGroup>
 
   <ItemGroup>
     <Content Include="JsonSerializationSpec.JsonApprovalSpecWithParent.approved.json">
@@ -20,12 +15,11 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Akka.TestKit.Xunit2" Version="1.3.9" />
-    <PackageReference Include="FluentAssertions" Version="5.2.0" />
-    <PackageReference Include="FsCheck.Xunit" Version="2.10.8" />
+    <PackageReference Include="Akka.TestKit.Xunit2" Version="$(AkkaVersion)" />
+    <PackageReference Include="FluentAssertions" Version="5.9.0" />
+    <PackageReference Include="FsCheck.Xunit" Version="2.14.0" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="$(TestSdkVersion)" />
     <PackageReference Include="xunit" Version="$(XunitVersion)" />
-    <DotNetCliToolReference Include="dotnet-xunit" Version="$(XunitVersion)" />
     <PackageReference Include="xunit.runner.visualstudio" Version="$(XunitVersion)" />
   </ItemGroup>
   
