@@ -1,15 +1,39 @@
 ﻿// -----------------------------------------------------------------------
 // <copyright file="B3Propagator.cs" company="Petabridge, LLC">
-//      Copyright (C) 2018 - 2018 Petabridge, LLC <https://petabridge.com>
+//      Copyright (C) 2015 - 2019 Petabridge, LLC <https://petabridge.com>
 // </copyright>
 // -----------------------------------------------------------------------
 
-using System.Globalization;
+using System.Linq;
 using OpenTracing.Propagation;
 using Petabridge.Tracing.Zipkin.Exceptions;
 
 namespace Petabridge.Tracing.Zipkin.Propagation
 {
+    /// <summary>
+    ///     Implements the "single header" B3 propagation format
+    /// </summary>
+    /// <remarks>
+    ///     See https://github.com/openzipkin/b3-propagation/issues/21 for full specification
+    /// </remarks>
+    public sealed class B3SingleHeaderPropagator : IPropagator<ITextMap>
+    {
+        internal const string B3SingleHeader = "b3";
+
+        public void Inject(SpanContext context, ITextMap carrier)
+        {
+            carrier.Set(B3SingleHeader, B3SingleHeaderFormatter.WriteB3SingleFormat(context));
+        }
+
+        public SpanContext Extract(ITextMap carrier)
+        {
+            var b3Entry = carrier.Any(x => x.Key.Equals(B3SingleHeader));
+            if (!b3Entry)
+                return null;
+            return B3SingleHeaderFormatter.ParseB3SingleFormat(carrier.Single(x => x.Key.Equals(B3SingleHeader)).Value);
+        }
+    }
+
     /// <inheritdoc />
     /// <summary>
     ///     Propagation system using B3 Headers supported by Zipkin
@@ -19,14 +43,42 @@ namespace Petabridge.Tracing.Zipkin.Propagation
     /// </remarks>
     public sealed class B3Propagator : IPropagator<ITextMap>
     {
-        internal const string B3TraceId = "X-B3-TraceId";
-        internal const string B3SpanId = "X-B3-SpanId";
-        internal const string B3ParentId = "X-B3-ParentSpanId";
-        internal const string B3Sampled = "X-B3-Sampled";
-        internal const string B3Debug = "X-B3-Flags";
+        internal const string B3TraceId = "x-b3-traceid";
+        internal const string B3SpanId = "x-b3-spanid";
+        internal const string B3ParentId = "x-b3-parentspanid";
+        internal const string B3Sampled = "x-b3-sampled";
+        internal const string B3Debug = "x-b3-flags";
+        private readonly B3SingleHeaderPropagator _singleHeaderPropagator = new B3SingleHeaderPropagator();
+        private readonly bool _useB3SingleHeader;
+
+        /// <inheritdoc />
+        /// <summary>
+        ///     Default constructor for the B3 Propagator. Doesn't use single-header format by default.
+        /// </summary>
+        public B3Propagator() : this(false)
+        {
+        }
+
+        /// <summary>
+        ///     Creates a new instance of the B3 Propagator.
+        /// </summary>
+        /// <param name="useB3SingleHeader">
+        ///     When set to <c>true</c>, enables the propagator to use the B3 single header
+        ///     propagation.
+        /// </param>
+        public B3Propagator(bool useB3SingleHeader)
+        {
+            _useB3SingleHeader = useB3SingleHeader;
+        }
 
         public void Inject(SpanContext context, ITextMap carrier)
         {
+            if (_useB3SingleHeader)
+            {
+                _singleHeaderPropagator.Inject(context, carrier);
+                return;
+            }
+
             carrier.Set(B3TraceId, context.TraceId);
             carrier.Set(B3SpanId, context.SpanId);
             if (context.ParentId != null)
@@ -49,6 +101,11 @@ namespace Petabridge.Tracing.Zipkin.Propagation
 
         public SpanContext Extract(ITextMap carrier)
         {
+            // try to extract the single B3 propagation value instead
+            var single = _singleHeaderPropagator.Extract(carrier);
+            if (single != null)
+                return single;
+
             TraceId? traceId = null;
             string spanId = null;
             string parentId = null;
@@ -56,7 +113,7 @@ namespace Petabridge.Tracing.Zipkin.Propagation
             var sampled = false;
             const bool shared = false;
             foreach (var entry in carrier)
-                switch (entry.Key)
+                switch (entry.Key.ToLowerInvariant())
                 {
                     case B3TraceId:
                         if (!TraceId.TryParse(entry.Value, out var t))
@@ -75,12 +132,13 @@ namespace Petabridge.Tracing.Zipkin.Propagation
                             debug = true;
                         break;
                     case B3Sampled:
-                        if (entry.Value.Equals("1"))
+                        if (entry.Value.Equals("1") || entry.Value.ToLowerInvariant().Equals("true")
+                        ) // support older tracers https://github.com/petabridge/Petabridge.Tracing.Zipkin/issues/72
                             sampled = true;
                         break;
                 }
 
-            if(traceId != null && spanId != null) // don't care of ParentId is null or not
+            if (traceId != null && spanId != null) // don't care of ParentId is null or not
                 return new SpanContext(traceId.Value, spanId, parentId, debug, sampled, shared);
             return null;
         }
