@@ -1,6 +1,6 @@
 ﻿// -----------------------------------------------------------------------
 // <copyright file="ZipkinFixture.cs" company="Petabridge, LLC">
-//      Copyright (C) 2018 - 2018 Petabridge, LLC <https://petabridge.com>
+//      Copyright (C) 2015 - 2019 Petabridge, LLC <https://petabridge.com>
 // </copyright>
 // -----------------------------------------------------------------------
 
@@ -17,13 +17,11 @@ namespace Petabridge.Tracing.Zipkin.Integration.Tests
 {
     public class ZipkinFixture : IAsyncLifetime
     {
-        private const string ZipkinImageName = "openzipkin/zipkin";
-        private readonly string _zipkinContainerName = $"zipkin-{Guid.NewGuid():N}";
-        private DockerClient _client;
+        protected const string ZipkinImageName = "openzipkin/zipkin";
+        protected readonly string _zipkinContainerName = $"zipkin-{Guid.NewGuid():N}";
+        protected DockerClient _client;
 
-        public string ZipkinUrl { get; private set; }
-
-        public async Task InitializeAsync()
+        public ZipkinFixture()
         {
             DockerClientConfiguration config;
             if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
@@ -34,7 +32,29 @@ namespace Petabridge.Tracing.Zipkin.Integration.Tests
                 throw new NotSupportedException($"Unsupported OS [{RuntimeInformation.OSDescription}]");
 
             _client = config.CreateClient();
+        }
 
+        public string ZipkinUrl { get; protected set; }
+
+        public virtual async Task InitializeAsync()
+        {
+            ZipkinUrl = await StartZipkinContainer(null, null);
+            await Task.Delay(TimeSpan.FromSeconds(20));
+        }
+
+        public virtual async Task DisposeAsync()
+        {
+            if (_client != null)
+            {
+                await _client.Containers.StopContainerAsync(_zipkinContainerName, new ContainerStopParameters());
+                await _client.Containers.RemoveContainerAsync(_zipkinContainerName,
+                    new ContainerRemoveParameters {Force = true});
+                _client.Dispose();
+            }
+        }
+
+        protected async Task<string> StartZipkinContainer(string[] links, string[] environmentArgs)
+        {
             var images = await _client.Images.ListImagesAsync(new ImagesListParameters {MatchName = ZipkinImageName});
             if (images.Count == 0)
                 await _client.Images.CreateImageAsync(
@@ -68,25 +88,15 @@ namespace Petabridge.Tracing.Zipkin.Integration.Tests
                                 }
                             }
                         }
-                    }
-                }
+                    },
+                    Links = links
+                },
+                Env = environmentArgs
             });
 
             // start the container
             await _client.Containers.StartContainerAsync(_zipkinContainerName, new ContainerStartParameters());
-            ZipkinUrl = $"localhost:{zipkinHttpPort}";
-            await Task.Delay(TimeSpan.FromSeconds(20));
-        }
-
-        public async Task DisposeAsync()
-        {
-            if (_client != null)
-            {
-                await _client.Containers.StopContainerAsync(_zipkinContainerName, new ContainerStopParameters());
-                await _client.Containers.RemoveContainerAsync(_zipkinContainerName,
-                    new ContainerRemoveParameters {Force = true});
-                _client.Dispose();
-            }
+            return $"localhost:{zipkinHttpPort}";
         }
     }
 }
