@@ -1,6 +1,6 @@
 ﻿// -----------------------------------------------------------------------
 // <copyright file="B3PropagatorSpecs.cs" company="Petabridge, LLC">
-//      Copyright (C) 2018 - 2018 Petabridge, LLC <https://petabridge.com>
+//      Copyright (C) 2015 - 2018 Petabridge, LLC <https://petabridge.com>
 // </copyright>
 // -----------------------------------------------------------------------
 
@@ -9,7 +9,9 @@ using FluentAssertions;
 using FsCheck;
 using FsCheck.Xunit;
 using OpenTracing.Propagation;
+using OpenTracing.Util;
 using Petabridge.Tracing.Zipkin.Propagation;
+using Petabridge.Tracing.Zipkin.Reporting.NoOp;
 using Petabridge.Tracing.Zipkin.Tracers;
 using Xunit;
 
@@ -17,12 +19,12 @@ namespace Petabridge.Tracing.Zipkin.Tests.Propagation
 {
     public class B3PropagatorSpecs
     {
-        public readonly MockZipkinTracer Tracer;
-
         public B3PropagatorSpecs()
         {
             Tracer = new MockZipkinTracer(propagtor: new B3Propagator());
         }
+
+        public readonly MockZipkinTracer Tracer;
 
         [Property(DisplayName = "Should be able to extract and inject spans via B3 headers")]
         public Property ShouldExtractAndInjectSpansViaB3(long traceIdHigh, long traceIdLow, long spanId, long? parentId,
@@ -59,7 +61,31 @@ namespace Petabridge.Tracing.Zipkin.Tests.Propagation
         }
 
         /// <summary>
-        /// Fix for https://github.com/petabridge/Petabridge.Tracing.Zipkin/issues/55
+        ///     Verify fix for https://github.com/petabridge/Petabridge.Tracing.Zipkin/issues/56
+        /// </summary>
+        [Fact(DisplayName =
+            "Bugfix for issue 56 - Propagator should not throw when attempting to inject non-Zipkin context.")]
+        public void BugFix56NonZipkinContextShouldNotThrowUponInjectionAttempt()
+        {
+            // uses the MockZipkinTracer
+            var carrier = new Dictionary<string, string>();
+            Tracer.Inject(NoOpZipkinSpanContext.Instance, BuiltinFormats.HttpHeaders,
+                new TextMapInjectAdapter(carrier));
+            carrier.Count.Should().Be(0);
+
+            // uses the real Zipkin tracer with a No-Op reporter
+            var testTracer = new ZipkinTracer(new ZipkinTracerOptions(
+                new Endpoint("test"), new NoOpReporter())
+            {
+                ScopeManager = new AsyncLocalScopeManager()
+            });
+            testTracer.Inject(NoOpZipkinSpanContext.Instance, BuiltinFormats.HttpHeaders,
+                new TextMapInjectAdapter(carrier));
+            carrier.Count.Should().Be(0);
+        }
+
+        /// <summary>
+        ///     Verify for https://github.com/petabridge/Petabridge.Tracing.Zipkin/issues/55
         /// </summary>
         [Fact(DisplayName = "B3Propagator should not extract SpanContext when none found")]
         public void ShouldNotExtractAnyTraceIdWhenNoneFound()
@@ -67,9 +93,130 @@ namespace Petabridge.Tracing.Zipkin.Tests.Propagation
             // pass in an empty carrier
             var carrier = new Dictionary<string, string>();
             var extracted =
-                (SpanContext)Tracer.Extract(BuiltinFormats.HttpHeaders, new TextMapExtractAdapter(carrier));
+                (SpanContext) Tracer.Extract(BuiltinFormats.HttpHeaders, new TextMapExtractAdapter(carrier));
 
             extracted.Should().BeNull();
+        }
+
+        /// <summary>
+        /// https://github.com/petabridge/Petabridge.Tracing.Zipkin/issues/71
+        /// </summary>
+        [Fact(DisplayName = "B3Propagator should extract headers without case sensitivity")]
+        public void B3HeaderExtractionShouldBeCaseInsensitive()
+        {
+            var carrier = new Dictionary<string, string>();
+
+            
+            var traceId = Tracer.IdProvider.NextTraceId();
+            var spanId = Tracer.IdProvider.NextSpanId();
+            var parentId = Tracer.IdProvider.NextSpanId();
+
+            // all upper case
+            carrier[B3Propagator.B3TraceId.ToUpperInvariant()] = traceId.ToString();
+
+            // all upper case
+            carrier[B3Propagator.B3SpanId.ToLowerInvariant()] = spanId;
+
+            // mixed case
+            carrier["X-B3-ParentSPANid"] = parentId;
+
+            var extracted =
+                (SpanContext)Tracer.Extract(BuiltinFormats.HttpHeaders, new TextMapExtractAdapter(carrier));
+
+            extracted.TraceId.Should().Be(traceId.ToString());
+            extracted.SpanId.Should().Be(spanId);
+            extracted.ParentId.Should().Be(parentId);
+        }
+
+        /// <summary>
+        /// https://github.com/petabridge/Petabridge.Tracing.Zipkin/issues/72
+        /// </summary>
+        [Fact(DisplayName = "B3Propagator should tolerate 'true' value for X-B3-Sampled field")]
+        public void B3HeaderExtractionShouldTolerateTrueForSampled()
+        {
+            var traceId = Tracer.IdProvider.NextTraceId();
+            
+            // sampled is set to `true`
+            var context = new SpanContext(traceId, Tracer.IdProvider.NextSpanId(), Tracer.IdProvider.NextSpanId(), false, true);
+            var carrier = new Dictionary<string, string>();
+
+            Tracer.Inject(context, BuiltinFormats.HttpHeaders, new TextMapInjectAdapter(carrier));
+
+            // validate true case
+            carrier[B3Propagator.B3Sampled] = "true";
+            var extracted =
+                (SpanContext)Tracer.Extract(BuiltinFormats.HttpHeaders, new TextMapExtractAdapter(carrier));
+
+            extracted.Sampled.Should().BeTrue();
+
+            // validate false case
+            carrier[B3Propagator.B3Sampled] = "false";
+            var extracted2 =
+                (SpanContext)Tracer.Extract(BuiltinFormats.HttpHeaders, new TextMapExtractAdapter(carrier));
+
+            extracted2.Sampled.Should().BeFalse();
+        }
+
+        /// <summary>
+        /// https://github.com/petabridge/Petabridge.Tracing.Zipkin/issues/66
+        /// </summary>
+        [Fact(DisplayName = "B3Propagator should be able to parse single header format")]
+        public void B3HeaderExtractionShouldHandleSingleHeaderFormat()
+        {
+            var traceId = Tracer.IdProvider.NextTraceId();
+
+            var context = new SpanContext(traceId, Tracer.IdProvider.NextSpanId(), Tracer.IdProvider.NextSpanId(), false, true);
+            var carrier = new Dictionary<string, string>();
+
+            var b3SingleHeader = new B3SingleHeaderPropagator();
+            b3SingleHeader.Inject(context, new TextMapInjectAdapter(carrier));
+            
+            var extracted =
+                (SpanContext)Tracer.Extract(BuiltinFormats.HttpHeaders, new TextMapExtractAdapter(carrier));
+
+            extracted.Should().Be(context);
+        }
+
+        /// <summary>
+        /// https://github.com/petabridge/Petabridge.Tracing.Zipkin/issues/66
+        /// </summary>
+        [Fact(DisplayName = "B3Propagator should be able to inject single header format")]
+        public void B3HeaderInjectionShouldHandleSingleHeaderFormat()
+        {
+            var traceId = Tracer.IdProvider.NextTraceId();
+
+            var context = new SpanContext(traceId, Tracer.IdProvider.NextSpanId(), Tracer.IdProvider.NextSpanId(), false, true);
+            var carrier = new Dictionary<string, string>();
+
+            // tracer2 will inject with single headers
+            var tracer2 = new MockZipkinTracer(propagtor:new B3Propagator(true));
+            tracer2.Inject(context, BuiltinFormats.HttpHeaders, new TextMapInjectAdapter(carrier));
+
+            // single header only
+            carrier.Count.Should().Be(1);
+
+            // tracer1 will still be able to read and extract single headers
+            var extracted = Tracer.Extract(BuiltinFormats.HttpHeaders, new TextMapExtractAdapter(carrier));
+            extracted.Should().Be(context);
+        }
+
+        [Fact(DisplayName = "Should be able to override IPropagator implementation via ZipkinTracerOptions")]
+        public void ShouldOverridePropagatorViaZipkinTracerOptions()
+        {
+            var tracerOptions = new ZipkinTracerOptions(Endpoint.Testing, new NoOpReporter())
+            {
+                Propagator = new B3SingleHeaderPropagator()
+            };
+
+            var tracer = new ZipkinTracer(tracerOptions);
+
+            var traceId = Tracer.IdProvider.NextTraceId();
+
+            var context = new SpanContext(traceId, Tracer.IdProvider.NextSpanId(), Tracer.IdProvider.NextSpanId(), false, true);
+            var carrier = new Dictionary<string, string>();
+
+            tracer.Inject(context, BuiltinFormats.HttpHeaders, new TextMapInjectAdapter(carrier));
+            carrier.Count.Should().Be(1);
         }
     }
 }
