@@ -1,6 +1,6 @@
 ﻿// -----------------------------------------------------------------------
 // <copyright file="ISpanIdProvider.cs" company="Petabridge, LLC">
-//      Copyright (C) 2018 - 2018 Petabridge, LLC <https://petabridge.com>
+//      Copyright (C) 2015 - 2019 Petabridge, LLC <https://petabridge.com>
 // </copyright>
 // -----------------------------------------------------------------------
 
@@ -23,7 +23,7 @@ namespace Petabridge.Tracing.Zipkin
         TraceId NextTraceId();
 
         /// <summary>
-        /// Generates a new unique span id.
+        ///     Generates a new unique span id.
         /// </summary>
         /// <returns>A new span id.</returns>
         string NextSpanId();
