@@ -1,9 +1,16 @@
 ﻿// -----------------------------------------------------------------------
 // <copyright file="Friends.cs" company="Petabridge, LLC">
-//      Copyright (C) 2018 - 2018 Petabridge, LLC <https://petabridge.com>
+//      Copyright (C) 2015 - 2019 Petabridge, LLC <https://petabridge.com>
 // </copyright>
 // -----------------------------------------------------------------------
 
 using System.Runtime.CompilerServices;
 
+/*
+[assembly: InternalsVisibleTo("Petabridge.Tracing.Zipkin.Tests, PublicKey=00240000048000009400000006020000002400005253413100040000010001000179e21e365117"+
+"e1758def0b71a5d19119ab26d983000a8d76cab1adce4e3f6482b006b2a4f6783a2d95fe31c972" +
+"5f1fe455bb88b0b98f37489a726cbf48918c7dea2d0fe96aeaaab5ea91767918325c0e576863f3" +
+"e72c10f4d61daee4886bc9d7fa2dd6039cea0bb935bb8fc103a3d387128e25ac5b29ed637b7cfa" +
+"acceaeb2")]
+*/
 [assembly: InternalsVisibleTo("Petabridge.Tracing.Zipkin.Tests")]
