@@ -1,6 +1,6 @@
 ﻿// -----------------------------------------------------------------------
 // <copyright file="ZipkinHttpApiTransmitter.cs" company="Petabridge, LLC">
-//      Copyright (C) 2018 - 2018 Petabridge, LLC <https://petabridge.com>
+//      Copyright (C) 2015 - 2019 Petabridge, LLC <https://petabridge.com>
 // </copyright>
 // -----------------------------------------------------------------------
 
@@ -9,7 +9,6 @@ using System.Collections.Generic;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.IO;
 
 namespace Petabridge.Tracing.Zipkin.Reporting.Http
 {
@@ -28,11 +27,6 @@ namespace Petabridge.Tracing.Zipkin.Reporting.Http
         /// </summary>
         public const string SpanPostUriPath = "api/v2/spans";
 
-        /// <summary>
-        ///     Only need one of these globally, and it's thread-safe. The streams it accesses internally are inherently not safe.
-        /// </summary>
-        private static readonly RecyclableMemoryStreamManager StreamManager = new RecyclableMemoryStreamManager();
-
         private readonly HttpClient _client;
 
         private readonly ISpanSerializer _serializer = new JsonSpanSerializer();
@@ -47,7 +41,8 @@ namespace Petabridge.Tracing.Zipkin.Reporting.Http
 
         public async Task<HttpResponseMessage> TransmitSpans(IEnumerable<Span> spans, TimeSpan timeout)
         {
-            using (var stream = StreamManager.GetStream("Petabridge.Tracing.Zipkin.HttpTransmitter"))
+            using (var stream =
+                SerializationStreamManager.StreamManager.GetStream("Petabridge.Tracing.Zipkin.HttpTransmitter"))
             {
                 _serializer.Serialize(stream, spans);
                 var cts = new CancellationTokenSource(timeout);
