@@ -1,16 +1,26 @@
-﻿namespace Petabridge.Tracing.Zipkin.Sampling
+﻿// -----------------------------------------------------------------------
+// <copyright file="ExternalSampling.cs" company="Petabridge, LLC">
+//      Copyright (C) 2015 - 2019 Petabridge, LLC <https://petabridge.com>
+// </copyright>
+// -----------------------------------------------------------------------
+
+namespace Petabridge.Tracing.Zipkin.Sampling
 {
     /// <inheritdoc />
     /// <summary>
-    /// Used when sampling is running, but it's driven by an external engine
-    /// rather than anything internal to this Zipkin driver itself.
+    ///     Used when sampling is running, but it's driven by an external engine
+    ///     rather than anything internal to this Zipkin driver itself.
     /// </summary>
     public sealed class ExternalSampling : ITraceSampler
     {
         public static readonly ExternalSampling Instance = new ExternalSampling();
-        private ExternalSampling() { }
+
+        private ExternalSampling()
+        {
+        }
 
         public bool Sampling { get; } = true;
+
         public bool IncludeInSample(string operationName)
         {
             /*
